animastery-ai/backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
langchain-core
langchain-community
langchain-openai
pydantic
numpy
chromadb
requests
pandas
python-multipart
networkx